Fix CI portability failures: clang sign-conversion, MSVC M_PI/posix_memalign, libmysofa CMP0077
Four CI jobs failed on the PR's first full-matrix run; three root causes:

- linux-clang / macos-appleclang: clang's -Wconversion implies
  -Wsign-conversion (GCC's does not), flagging int indexing in max_re.h.
  Reproduced locally with clang 18 + AMBITAP_WERROR=ON; these were the only
  two instances tree-wide. Fixed with explicit size_t casts.

- windows-msvc: tests still used M_PI (only the library headers had been
  purged; MSVC hides it behind _USE_MATH_DEFINES) -> replaced with
  ambitap::k_pi in all five test files. test_rt_safety.cpp used
  posix_memalign, which Windows lacks -> portable
  aligned_alloc_portable/_aligned_free wrappers, with the aligned delete
  overloads routed through the aligned free path (required on Windows,
  where _aligned_malloc memory must not go to plain free).

- sofa leg: libmysofa's pre-3.10 CMake leaves CMP0077 unset, so its
  option(BUILD_TESTS) CLEARED the plain normal variable and its CUnit test
  dependency was required again. Default the policy to NEW for the fetched
  subtree (CMAKE_POLICY_DEFAULT_CMP0077) so the variable is honored without
  clobbering the consumer's cache.

Verified locally: gcc and clang 18 builds with AMBITAP_WERROR=ON are
warning-free, 100 tests pass on both plus ASan+UBSan, and the format gate
is clean.

+++ include/ambitap/math/decoding/max_re.h
@@ -27,7 +27,7 @@ namespace ambitap {
         const float theta = (137.9f * k_pi / 180.0f) / (static_cast<float>(order) + 1.51f);
         const float x     = std::cos(theta);
 
-        std::vector<float> weights(order + 1);
+        std::vector<float> weights(static_cast<size_t>(order) + 1);
 
         // Legendre polynomials via recurrence: P_0(x)=1, P_1(x)=x,
         // (n+1) P_{n+1}(x) = (2n+1) x P_n(x) - n P_{n-1}(x)
@@ -42,9 +42,9 @@ namespace ambitap {
             float p_next =
                 (static_cast<float>(2 * n + 1) * x * p_curr - static_cast<float>(n) * p_prev)
                 / static_cast<float>(n + 1);
-            p_prev         = p_curr;
-            p_curr         = p_next;
-            weights[n + 1] = p_next;
+            p_prev                              = p_curr;
+            p_curr                              = p_next;
+            weights[static_cast<size_t>(n) + 1] = p_next;
         }
 
         return weights;
